feat: style navbar
Loaded a new font for the website for a more polished and distinct look. Applied background color to the page and set a width to the content to improve readability.
Applied the brand color to the navbar and arranged and centered all items horizontally to achieve the traditional navbar-look. Separated the menu from the logo.
Rounded the corners of the logo, with a circle format the cauldron is difficult to see. Changed color and size of the font for branding and better visual appeal.