feat: style the gallery
Applied white background in accordance with the brand colors to the section and centered the content in alignment with previous sections.
Styled to font using brand colors and font. Created a rounded border around the icons to emphasize the "special effects" that will be added at a later time.
Added padding where needed for better visual appeal.