feat: style quote and testimonial section
Originally this was supposed to be one section but was separated into two to distinguish the quote and to make the font of the testimonials look better.

A background color was applied to the quote section and the quote was centered. The font was styled to match the branding and the source was aligned where it
would be expected.

A white background was applied to the testimonial section to frame the quote together with the gallery section. The testimonials have been positioned in place,
the font adjusted to fit branding and mirror the gallery section. The images were rounded.